Do not remove test directory configured with NODE_FB_TEST_TMP_DIR

+++ packages/node-firebird-driver/src/test/tests.ts
@@ -106,7 +106,7 @@ export function runCommonTests(client: Client) {
 
       expect(client.isValid).toBeFalsy();
 
-      if (isLocal()) {
+      if (isLocal() && !testConfig.tmpDir) {
         fs.rmdirSync(testConfig.tmpDir!);
       }
     });
